llm-backend/requirements.txt
fastapi
uvicorn[standard]
celery
redis
llama-cpp-python
pydantic
instructor
python-dotenv
